prettify rank percentile
